chore: util function
