+++ src/util/monitorUtils.ts
@@ -42,3 +42,48 @@ export const getDeparturesWithoutHiddenRoutes = (
   });
   return departures;
 };
+
+export const createDepartureArray = (views, stops, isStation = false) => {
+  const defaultSettings = {
+    hiddenRoutes: [],
+    timeshift: 0,
+  }
+  const departures = [];
+  const stringsToTranslate = [];
+  views.forEach((view, i) => {
+    Object.keys(view.columns).forEach(column => {
+      const departureArray = [];
+      stops.forEach(stop => {
+        const stopIndex = view.columns[column].stops
+          .map(stop => stop.gtfsId)
+          .indexOf(stop.gtfsId);
+        if (stopIndex >= 0) {
+          if (isStation) {
+          stop.stops.forEach(s => {
+                      s.routes.forEach(r => {
+                        stringsToTranslate.push(...r.patterns.map(p => p.headsign));
+                      });
+                    });
+      } else {
+          stop.patterns.forEach(r => {
+            stringsToTranslate.push(r.headsign);
+          });
+        }
+          const { hiddenRoutes, timeshift } =
+            view.columns[column].stops[stopIndex].settings ? view.columns[column].stops[stopIndex].settings : defaultSettings;
+          departureArray.push(
+            ...getDeparturesWithoutHiddenRoutes(
+              stop,
+              hiddenRoutes,
+              timeshift,
+            ),
+          );
+        }
+      });
+      const colIndex = column === 'left' ? 0 : 1;
+      departures[i] = departures[i] ? departures[i] : [[], []];
+      departures[i][colIndex] = departureArray;
+    });
+  })
+  return [stringsToTranslate, departures];
+};
